==> modeling/toc/04_apply_TOC_model.Rmd <==
---
title: "Applying TOC preds on sensor data"
author: "Sam Struthers-ROSSyndicate"
date: "2025-08-01"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
source("src/setup_libraries.R")
#source("src/TOC_model_train.R")
library(purrr)
library(ggplot2)
library(dplyr)
library(gridExtra)
library(grid)
library(Ckmeans.1d.dp)
library(DiagrammeR)
library(Metrics)
library(caret)
library(glue)
set.seed(123)


sites <- c("chd", "joei", "cbri", "sfm", "penn", "lbea", "pfal", "pbr", "pman", "pbd", "udall")
features <- c('FDOM', 'Sensor_Turb', "log_sensor_turb", 'Sensor_Cond', 'Chl_a','Temp','f_c_turb', "log_chl_a" )
target  = "TOC"
```


Run 03 script to create model then apply to longer/full dataset to see results


# Setup

```{r}
# Load saved scaling parameters
scaling_params <- readRDS("data/upper_clp_dss/modeling/scaling_parameters_20250820.rds")

# Function to apply the same scaling to new data
apply_training_scale <- function(new_data, scaling_params) {
  numeric_cols <- names(select(new_data, where(is.numeric)))
  
  scaled_data <- new_data
  
  for (col in numeric_cols) {
    min_val <- scaling_params[[paste0(col, "_min")]]
    max_val <- scaling_params[[paste0(col, "_max")]]
    # Apply same min-max scaling as training data
    scaled_data[[col]] <- (new_data[[col]] - min_val) / (max_val - min_val)
  }
  
  return(scaled_data)
}
```

# Read in water chem data
```{r}
water_chem <- read_parquet("data/upper_clp_dss/modeling/ROSS_FC_water_chemistry.parquet")%>%
  mutate(DT_round = round_date(DT_sample, unit = "1 hour" ))%>%
  rename(site = site_code)#%>%
#filter(collector == "ROSS")

toc_max <- max(water_chem$TOC, na.rm = T)

```

# read in sensor data and format to match model needs


```{r}
#generated in 01 script (all_sensor_data_wide)

prepped_data <- read_parquet("data/upper_clp_dss/sensor/prepped/all_sensor_data_2023-03-30_2025-08-27.parquet")%>%
  select(site, 
         DT_round,
         FDOM = `FDOM Fluorescence`,
         #hrs_since_last_cleaning = hours_since_last_visit,
         Temp = Temperature,
         Sensor_Cond = `Specific Conductivity`,
         Sensor_Turb = Turbidity,
         Chl_a = `Chl-a Fluorescence`)%>%
  filter(site %in% sites)%>% # just grab sites of interest
  mutate(#capping sensor turb at 1000 for outliers
    Sensor_Turb = ifelse(Sensor_Turb > 1000, 1000,
                         ifelse(Sensor_Turb <= 0, 0.01, Sensor_Turb)), # if turb is zero, replace with 0.01 to avoid division by zero
    log_sensor_turb = log(Sensor_Turb), 
    Chl_a = ifelse(Chl_a == 0,0.0001, Chl_a),# if Chl_A is zero, replace with 0.001 to avoid division by zero
    log_chl_a = log(Chl_a),
    #compute optical bands
    f_c_turb = FDOM/(Chl_a + Sensor_Turb + FDOM))%>%
  #remove NAs
  filter( !is.na(f_c_turb) & !is.na(Temp)& !is.na(Sensor_Cond))#!is.na(hrs_since_last_cleaning) 

# Apply to your new data
prepped_data_norm <- prepped_data %>%
  #normalize data based on full sensor set from 01_sensor_data_prep.rmd
  apply_training_scale(., scaling_params)%>%
  na.omit()
```


# Run model on sensor data to get TOC estimate

```{r}
most_recent_file <- list.files("data/upper_clp_dss/modeling/model_splits", pattern = "toc_xgboost_models_light_.*\\.rds", full.names = TRUE) %>%
  tail(1)
model <- readRDS(file = most_recent_file)
```

# Make a plot of historical data model estimated TOC

```{r}

create_toc_model_plot <- function(summary_interval, site_sel,site_title,start_DT,end_DT){
  
  #convert datetimes  
  start_DT <- ymd_hm(start_DT,tz = "MST")
  end_DT <- ymd_hm(end_DT,tz = "MST")
  # Get testing data RMSE for the model
  # find half way point between start and end DT
  halfway_DT <- start_DT + (end_DT - start_DT) / 2
  
  # Summarize data to desired interval and trim to individual site
  sum_data_norm <- prepped_data_norm%>%
    filter(site == site_sel)%>%
    select(site, DT_round, all_of(features)) %>%
    mutate(DT_round = round_date(DT_round, unit = summary_interval )) %>% # ensure DT_round is rounded to the nearest
    group_by(site, DT_round) %>%
    summarise(across(everything(), mean, na.rm = TRUE), .groups = 'drop')
  #convert to matrix
  model_input_matrix_sum <- sum_data_norm[, features]%>%
    mutate(across(everything(), as.numeric)) %>%
    as.matrix()
  
  # Add predictions from each fold model as new columns
  for (i in seq_along(model)) {
    fold_model <- model[[i]]
    col_name <- glue("{target}_guess_fold{i}")
    sum_data_norm[[col_name]] <- round(predict(fold_model, model_input_matrix_sum, iteration_range = c(1, fold_model$best_iteration)), 2)
  }
  
  fold_cols <- glue("{target}_guess_fold{seq_along(model)}")
  ensemble_col <- glue("{target}_guess_ensemble")
  sum_data_norm[[ensemble_col]] <- round(rowMeans(sum_data_norm[, fold_cols]), 2)

plot_sensor_data <- sum_data_norm %>%
  # Filter to desired time window first
  filter(between(DT_round, start_DT, end_DT)) %>%
  # Pad missing timestamps based on summary_interval
  pad(
    by = "DT_round",
    interval = summary_interval,   # e.g., "1 hour", "1 day", etc.
    group = NULL                   # optional: group by site if multiple sites
  ) %>%
  # Compute min/max across folds
  mutate(
    !!glue("{target}_guess_min") := pmin(!!!syms(fold_cols), na.rm = TRUE),
    !!glue("{target}_guess_max") := pmax(!!!syms(fold_cols), na.rm = TRUE),
    !!glue("{target}_guess_ensemble") := pmax(0, !!sym(glue("{target}_guess_ensemble"))), 
    group = with(rle(!is.na(.data[[ensemble_col]])), rep(seq_along(values), lengths))
    )

  
  #Water chem
  plot_water_chem <- water_chem %>%
    filter(site == site_sel)%>%
    filter(between(DT_round, start_DT, end_DT))%>%
    mutate(DT_round = round_date(DT_round, unit = summary_interval )) #round date to match sensor data
  
  #get correct dates for plot title
  start_DT <- min(plot_sensor_data$DT_round, na.rm = TRUE)
  end_DT <- max(plot_sensor_data$DT_round, na.rm = TRUE)


  #Create plot
 toc_plot<- ggplot() +
    geom_ribbon(data = plot_sensor_data,
                aes(x = DT_round,
                    ymin = TOC_guess_min,
                    ymax = TOC_guess_max,
                    fill = "Models Estimate Range"),
                alpha = 0.5) +
    # one line per fold_col, each with its own color
    # map(fold_cols, ~
    #   geom_path(data = plot_sensor_data,
    #             aes(x = DT_round,
    #                 y = .data[[.x]],
    #                 color = .x),    # <- put fold name in legend
    #             linewidth = 0.3)
    # ) +
  geom_line(
    data = plot_sensor_data,
    aes(x = DT_round,
        y = .data[[ensemble_col]],
        group = group,
        color = "Mean Model Estimate"),
    linewidth = 1
  ) +
    geom_point(data = plot_water_chem,
               aes(x = DT_round,
                   y = round(TOC,2),
                   #shape = collector,
                   color = "Sample Values")) +
    labs(title = paste0("Estimated TOC at ", site_title," from ", as.Date(start_DT), " - ", as.Date(end_DT)),
         x = "Date",
         y = "Model Estimated TOC (mg/L)", 
         # shape = "Sample Collector",
         fill = "",
         color = "") +
    scale_fill_manual(values = c("Models Estimate Range" = "grey")) +
    scale_color_manual(
      values = c(
        #setNames(scales::hue_pal()(length(fold_cols)), fold_cols),  # distinct colors for folds
        "Mean Model Estimate" = "#E70870",
        "Sample Values" = "#002EA3"
      )
    ) +
    theme(legend.position = "bottom") +
    guides(fill = guide_legend(override.aes = list(alpha = 0.5))) +
    annotate("text", 
             x = max(plot_sensor_data$DT_round, na.rm = TRUE), 
             y = Inf, 
             label = "PRELIMINARY RESULTS", 
             hjust = 1.1, 
             vjust = 1.5, 
             size = 7, 
             fontface = "bold",
             color = "red") +
    ROSS_theme
  
  # combine grabs and sensor data for plotting to determine axes
  plot_toc <- c(plot_sensor_data$TOC_guess_min, plot_sensor_data$TOC_guess_max, plot_water_chem$TOC)
  max_plot_toc <- max(plot_toc, na.rm = TRUE)
  min_plot_toc <- min(plot_toc, na.rm = TRUE)
  
  if(max_plot_toc > max(water_chem$TOC, na.rm = T)){
    toc_plot <- toc_plot + 
      geom_hline(yintercept = toc_max, linetype = "dashed", color = "red") +
      annotate("text", x = halfway_DT, y = toc_max - 0.2, label = "Max TOC Measured")
  }
  
  return(toc_plot)
  
}
```

## Create plots of the other sensor data

```{r}
create_sensor_ts_plot <- function(example_site, site_title, start_dt, end_dt){

example_data <- prepped_data %>%
  filter(site == example_site) %>%
  filter(between(DT_round, start_dt, end_dt)) %>%
  mutate(date = as.Date(DT_round)) %>%
  pivot_longer(
    cols = c("FDOM":"Chl_a"),
    names_to = "parameter",
    values_to = "value"
  ) %>%
  mutate(parameter = case_when(
    parameter == "FDOM" ~ "FDOM Fluorescence",
    parameter == "Sensor_Turb" ~ "Turbidity",
    parameter == "Sensor_Cond" ~ "Specific Conductivity",
    parameter == "Chl_a" ~ "Chl-a Fluorescence",
    parameter == "Temp" ~ "Temperature",
    TRUE ~ parameter
  )) %>%
  group_by(parameter) %>%       # pad separately for each parameter
  pad(
    by = "DT_round",
    interval = "15 min"
  ) %>%
  ungroup()

  
  example_grab <- water_chem%>%
    filter(site == example_site)%>%
    filter(between(DT_sample, start_dt, end_dt))

 example_ts_grab <- ggplot(example_data, aes(x = DT_round, y = value)) +
  geom_line(color = "#E70870", na.rm = TRUE) +  # breaks lines at NAs
  geom_vline(data = example_grab, aes(xintercept = DT_sample), 
             linetype = "dashed", color = "#002EA3") +
  facet_wrap(~parameter, scales = "free_y", ncol = 1) +
  labs(
    title = paste0("Sensor Time Series at Site: ", site_title),
    x = "Date",
    y = "Sensor Value"
  ) +
  ROSS_theme

  
  return(example_ts_grab)
}

```




# Example 1: CHD 2025

```{r}

chd_plot <- create_toc_model_plot(summary_interval = "1 hour",
                      site_sel = "chd",
                      site_title = "Chambers Lake Outflow",
                      start_DT = "2025-05-01 23:59",
                      end_DT = "2025-07-05 00:00"
)


ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/chambers_outflow_20250501_20250705_toc.png", 
       plot = chd_plot, 
       width = 12, 
       height = 8, 
       dpi = 300)

chd_ts_plot <- create_sensor_ts_plot(example_site = "chd",
                                   site_title = "Chambers Lake Outflow",
                                   start_dt = "2025-05-01 00:00",
                                   end_dt = "2025-07-05 00:00"
)


ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/chambers_outflow_20250501_20250705_sensor.png",
       plot = chd_ts_plot, 
       width = 10, 
       height = 10, 
       dpi = 300)

```


# Example 2: PFAL 2025

```{r}

pfal_plot <- create_toc_model_plot(site_sel = "pfal",
                                   summary_interval = "1 hours",
                                   site_title = "CLP @ Poudre Falls",
                                   start_DT = "2025-04-15 23:59",
                                   end_DT = "2025-07-10 00:00"
)

ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/pfal_20240415_20240710_toc.png", 
       plot = pfal_plot, 
       width = 12, 
       height = 8, 
       dpi = 300)

pfal_ts_plot <- create_sensor_ts_plot(example_site = "pfal",
                                   site_title = "CLP @ Poudre Falls",
                                   start_dt = "2025-04-15 00:00",
                                   end_dt = "2025-07-08 00:00"
)
ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/pfal_20240415_20240710_sensor.png",
       plot = pfal_ts_plot, 
       width = 10, 
       height = 10, 
       dpi = 300)

```

# Example 3: PBR

```{r}

pbr_plot <-  create_toc_model_plot(summary_interval = "1 hour",
                                   site_sel = "pbr",
                                   site_title = " CLP @ Indian Meadows",
                                   start_DT = "2025-05-01 00:00",
                                   end_DT = "2025-07-08 00:00"
)

ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/pbr_20250501_20250708_toc.png", 
       plot = pbr_plot, 
       width = 12, 
       height = 8, 
       dpi = 300)

pbr_ts_plot <- create_sensor_ts_plot(example_site = "pbr",
                                   site_title = "CLP @ Indian Meadows",
                                   start_dt = "2025-05-01 00:00",
                                   end_dt = "2025-07-08 00:00")

ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/pbr_20250501_20250708_sensor.png",
       plot = pbr_ts_plot, 
       width = 10, 
       height = 10, 
       dpi = 300)

```

# Example 4: SFM 2025

```{r}

sfm_plot<- create_toc_model_plot(site_sel = "sfm",
                                  summary_interval = "1 hour",
                                  site_title = "South Fork CLP",
                                  start_DT = "2025-04-15 23:59",
                                  end_DT = "2025-06-15 00:00"
)
sfm_plot

ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/sfm_20250424_20250615_toc.png", 
       plot = sfm_plot, 
       width = 12, 
       height = 8, 
       dpi = 300)

sfm_ts_plot <- create_sensor_ts_plot(example_site = "sfm",
                                   site_title = "South Fork CLP",
                                   start_dt = "2025-04-15 00:00",
                                   end_dt = "2025-06-15 00:00")
sfm_ts_plot
ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/sfm_20250415_20250615_sensor.png",
       plot = sfm_ts_plot, 
       width = 10, 
       height = 10, 
       dpi = 300)

```



# Example 4: PMAN 2025

```{r}

pman_plot<- create_toc_model_plot(site_sel = "pman",
                                  summary_interval = "1 hour",
                                  site_title = "Poudre at Manner's Bridge",
                                  start_DT = "2025-04-15 23:59",
                                  end_DT = "2025-07-05 00:00"
)
pman_plot

ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/pman_20250424_20250714_toc.png", 
       plot = pman_plot, 
       width = 12, 
       height = 8, 
       dpi = 300)
pman_ts_plot <- create_sensor_ts_plot(example_site = "pman",
                                   site_title = "Poudre at Manner's Bridge",
                                   start_dt = "2025-04-15 00:00",
                                   end_dt = "2025-07-05 00:00")
pman_ts_plot
ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/pman_20250415_20250705_sensor.png",
       plot = pman_ts_plot, 
       width = 10, 
       height = 10, 
       dpi = 300)

```

# Example 5: PBD 2025

```{r}

pbd_plot<- create_toc_model_plot(site_sel = "pbd",
                                 summary_interval = "1 hour",
                                 site_title = "Poudre at Canyon Mouth Gage",
                                 start_DT = "2025-05-20 23:59",
                                 end_DT = "2025-07-05 00:00"
)

pbd_plot

ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/pbd_20250415_20250705_toc.png", 
       plot = pbd_plot, 
       width = 12, 
       height = 8, 
       dpi = 300)

pbd_ts_plot <- create_sensor_ts_plot(example_site = "pbd",
                                   site_title = "Poudre at Canyon Mouth Gage",
                                   start_dt = "2025-05-20 00:00",
                                   end_dt = "2025-07-05 00:00")

ggsave("data/upper_clp_dss/figures/toc_modeling/timeseries/pbd_20250520_20250705_sensor.png",
       plot = pbd_ts_plot, 
       width = 10, 
       height = 10, 
       dpi = 300)

```

# Run all 
```{r}

```
